Do not overwrite keyRing element with itself
Using KeePassXC 2.5.0 with a database created with KeePassXC 2.5.0,
decrypted messages from keepassxc-proxy contain an 'oldHash' and a
'hash' value that are equal. In that case, we must not try to replace
the database hash ID in the keyRing with itself, as it effectively leads
to the keyRing entry to be destroyed without replacement when using the
current implementation.