Increase reliability of custom login fields functionality
Some (possibly a lot of) websites, like https://042564560466.signin.aws.amazon.com/console nowadays are generating dynamic classNames in their forms. ClassNames there depend on multitude factors, such as which field is currently selected or if field is empty. In case of formerly mentioned websites, this looks to be the case for AngularJS framework, which injects classes like `ng-untouched`, `ng-valid`, `ng-not-empty` and so on. This means keepassxc detects these fields first time only (and sometimes again randomly). I don't know reasons why this extension went for a solution where classNames were playing major part of the role of identifying the fields, but it turned out to be unreliable. This patch removes this factor and instead prefers to rely on more stable attributes, especially input/target name. Extra/custom field associations in other password field managers like 1Password7 and Enpass detect the fields automatically, very reliably. I think this is closer to how they do it.